fix: respect dataform local scope option
With this change, the plugin should respect the local scope option in
all cases. Previously, it was always showing an error when initializing
the extension, even though dataform was installed locally.
